+++ src/init.c
@@ -9,6 +9,7 @@ extern SEXP R_isValidJSON(SEXP);
 extern SEXP R_jsonPrettyPrint(SEXP, SEXP);
 extern SEXP R_json_parser_init_from_con(SEXP, SEXP, SEXP, SEXP, SEXP);
 extern SEXP R_readFromJSON(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
+extern SEXP R_toJSONAtomicFast(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
 
 
 static const R_CallMethodDef CallEntries[] = {
@@ -17,6 +18,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"R_jsonPrettyPrint",           (DL_FUNC) &R_jsonPrettyPrint,           2},
     {"R_json_parser_init_from_con", (DL_FUNC) &R_json_parser_init_from_con, 5},
     {"R_readFromJSON",              (DL_FUNC) &R_readFromJSON,              6},
+    {"R_toJSONAtomicFast",          (DL_FUNC) &R_toJSONAtomicFast,          6},
     {NULL, NULL, 0}
 };